+++ qtgql/gqlwstransport/operationhandler.hpp
@@ -1,14 +1,15 @@
 #pragma once
 #include <QObject>
 
-#include "../../../../bases/environment.hpp"
-#include "../../../../bases/metadata.hpp"
 #include "gqlwstransport.hpp"
+#include "qtgql/bases/bases.hpp"
 
 namespace qtgql {
+namespace gqlwstransport {
 
 class _OperationHandlerABCSignals : public QObject {
   Q_OBJECT
+
   Q_PROPERTY(bool completed READ completed NOTIFY completedChanged)
   Q_PROPERTY(bool operationOnFlight READ operation_on_flight NOTIFY
                  operationOnFlightChanged)
@@ -18,18 +19,24 @@ class _OperationHandlerABCSignals : public QObject {
   bool m_operation_on_the_fly = false;
 
  signals:
+
   void completedChanged();
+
   void operationOnFlightChanged();
+
   void error(const QJsonArray &);
 
  protected slots:
+
   void set_completed(bool v);
+
   void set_operation_on_flight(bool v);
 
  public:
   using QObject::QObject;
 
   bool completed() const;
+
   bool operation_on_flight();
 };
 
@@ -64,18 +71,23 @@ class OperationHandlerABC
       environment()->execute(a);
     }
   }
+
   void refetch() {
     set_completed(false);
     fetch();
   };
+
   const bases::HashAbleABC &message() override {
     m_message_template.set_variables(m_variables);
     return m_message_template;
   };
+
   void on_completed() override { set_completed(true); };
+
   void on_error(const QJsonArray &errors) override {
     set_completed(true);
     emit error(errors);
   };
 };
-};  // namespace qtgql
+};  // namespace gqlwstransport
+}  // namespace qtgql
