+++ qtgql/utils/utils.hpp
@@ -2,7 +2,19 @@
 #include <stdio.h>
 
 #include <QAbstractListModel>
+#include <QRegularExpression>
 namespace qtgql {
+namespace utils {
+
+inline std::optional<QString> get_operation_name(const QString& query) {
+  static QRegularExpression re("(subscription|mutation|query)( [0-9a-zA-Z]+)*");
+  auto match = re.match(query);
+  if (match.hasMatch()) {
+    return match.captured(2).trimmed();
+  }
+  return {};
+}
+
 class NotImplementedError : public std::logic_error {
   struct Msg {
     const char* msg = "Function not yet implemented";
@@ -11,4 +23,6 @@ class NotImplementedError : public std::logic_error {
  public:
   explicit NotImplementedError(const Msg& msg) : std::logic_error(msg.msg){};
 };
+}  // namespace utils
+
 }  // namespace qtgql
