+++ tests/test_codegen/generated_test_projects/DecimalTestCase/test_decimaltestcase.cpp
@@ -14,7 +14,8 @@ TEST_CASE("DecimalTestCase", "[generated-testcase]") {
   client->wait_for_valid();
 
   bases::Environment::set_gql_env(std::make_shared<bases::Environment>(
-      "DecimalTestCase", std::unique_ptr<qtgql::GqlWsTransportClient>(client)));
+      "DecimalTestCase",
+      std::unique_ptr<gqlwstransport::GqlWsTransportClient>(client)));
 
   auto mq = std::make_shared<mainquery::MainQuery>();
   mq->fetch();
