+++ tests/test_codegen/generated_test_projects/OptionalScalarsTestCase/test_optionalscalarstestcase.cpp
@@ -14,7 +14,7 @@ TEST_CASE("OptionalScalarsTestCase", "[generated-testcase]") {
 
   bases::Environment::set_gql_env(std::make_shared<bases::Environment>(
       "OptionalScalarsTestCase",
-      std::unique_ptr<qtgql::GqlWsTransportClient>(client)));
+      std::unique_ptr<gqlwstransport::GqlWsTransportClient>(client)));
   auto mq = std::make_shared<mainquery::MainQuery>();
 
   SECTION("when null returns default values") {
